Skip integration test, start simpler keeper test
